requirements.txt
yacs==0.1.8
numpy==1.23.4
pandas==2.0.1
matplotlib==3.6.2
scikit-image==0.20.0
scikit-learn==1.2.2
seaborn==0.12.2
comet_ml==3.33.1
python-dotenv==1.0.0
pybids==0.16.1
nilearn==0.10.1
scipy==1.9.1
nibabel==5.1.0
networkx==3.1
tqdm==4.65.0
h5py==3.8.0
pybids==0.16.1
ipykernel==6.23.0
fastdtw==0.3.4
PyWavelets==1.4.1
